requirements.txt
requests~=2.27.1
pandas~=1.4.2
python-dateutil~=2.8.2
pytz~=2022.1
Flask~=2.1.2
WTForms~=3.0.1
gunicorn==20.1.0